refactor: create gameDataHandler module
A file in which the game's data is stored was created. This file will
store the player's Gameboard object, the computer's Gameboard object,
and other related data. Additionally, the AllowPlayerShipPlacement
function has been moved to this file because it requires modifying
the player's Gameboard object when a ship is placed.